+++ mopidy/utils/__init__.py
@@ -5,6 +5,8 @@ import sys
 
 logger = logging.getLogger('mopidy.utils')
 
+
+# TODO: user itertools.chain.from_iterable(the_list)?
 def flatten(the_list):
     result = []
     for element in the_list:
@@ -14,22 +16,25 @@ def flatten(the_list):
             result.append(element)
     return result
 
+
 def import_module(name):
     __import__(name)
     return sys.modules[name]
+
 
 def get_class(name):
     logger.debug('Loading: %s', name)
     if '.' not in name:
         raise ImportError("Couldn't load: %s" % name)
     module_name = name[:name.rindex('.')]
-    class_name = name[name.rindex('.') + 1:]
+    cls_name = name[name.rindex('.') + 1:]
     try:
         module = import_module(module_name)
-        class_object = getattr(module, class_name)
+        cls = getattr(module, cls_name)
     except (ImportError, AttributeError):
         raise ImportError("Couldn't load: %s" % name)
-    return class_object
+    return cls
+
 
 def locale_decode(bytestr):
     try: